example/chatkit-backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
openai>=2.8.1
openai-agents>=0.6.1
openai-chatkit==1.1.2
python-dotenv>=1.0.0
pydantic>=2.9.0
httpx>=0.27.0
firebase-admin>=6.5.0
anthropic[bedrock]>=0.39.0
boto3>=1.35.0
claude-agent-sdk>=0.1.0
google-generativeai>=0.8.0
python-telegram-bot>=21.0
